+++ CHANGELOG.rst
@@ -19,6 +19,8 @@ Fixed
 
 - **General**
     - ``django-autocomplete-light`` Docker build crash with ``whitenoise`` (#1666)
+- **Samplesheets**
+    - Multi-file upload not working (#
 
 
 v0.13.3 (2023-05-10)